Lua file refactor
